pyproject.toml
[project]
name = "blip2-opt-2-7b-onnx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.9.1",
    "torchvision==0.24.1",
    "transformers==4.55.4",
    "safetensors==0.6.2",
    "huggingface_hub==0.36.0",
    "onnx==1.19.1",
    "onnxruntime==1.23.2",
    "onnxsim==0.4.36",
    "optimum[onnxruntime]==2.0.0",
    "onnxscript==0.5.6",
    "pillow==12.0.0",
    "accelerate==1.11.0",
    "bitsandbytes==0.48.2",
    "tritonclient[http]>=2.40.0",
    "requests>=2.31.0"
]
